Updated fim

+++ src/sieasymptotic/geom/geometry.py
@@ -6,7 +6,7 @@ def transform_cartesian_to_polar(x, y, omegatilde=0):
     Args:
         x (jnp.array): The x-coordinate of the position.
         y (jnp.array): The y-coordinate of the position.
-        omegatilde (int, optional): The angle between the major axis of the lens and the x-axis. Defaults to 0.
+        omegatilde (float, optional): The angle between the major axis of the lens and the x-axis. Defaults to 0.
 
     Returns:
         jnp.array: The radial and angular coordinates of the position.
